refactor: test의 common 패키지는 support 패키지로 통합
