+++ backend/src/test/java/com/woowacourse/kkogkkog/support/setup/ApplicationTest.java
@@ -1,6 +1,5 @@
-package com.woowacourse.kkogkkog.common.annotaion;
+package com.woowacourse.kkogkkog.support.setup;
 
-import com.woowacourse.kkogkkog.support.DataClearExtension;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
